digest max_na_sequence
This changes the implementation of `digest` to optionally handle data gaps. I think this'll be better than `trim_na_observations` in most (all?) cases, because long gaps that span two fragments can be filtered out, while short gaps that happen to be on the end of a fragment can be retained.

This necessitated also adding an ".observations" argument to `digest`, which is annoying.