digest adjusts datetimes
Incorporated an adjustment to datetimes to align them with the "now" data, in the same fashion as the `adj_idx` column.

This will allow all of that datetime finding portion of 1anneal` to be removed. That code existed as a byproduct of the `resolution` parameter, which now that I plan on removing it, can be made much simpler (i.e this).