anneal removed datetime calcs, upsampling
This greatly simplifies the annealing process. I can just use the `adj_datetime` from the digest and shift it the same as the `adj_idx`.

Also, since `resolution` is gone, can get rid of upsampling stuff.

Removing upsampling means that the `loess_fit` is no longer needed -- no more interpolation! Just use the fitted values contained in the digest's fragments.

This means that I now need to ask for the name of the fitted column (assumed identical b/w the original data and the digest... who would change it?), and use *that* with some dplyr programming magic in the loss_fn call.

I've also had to change the call signatures of the plotting methods. I'm not sure I'm happy with it, but it works for now.

I'm bizarrely concerned about a bug somewhere in these changes, but the plot snapshots are the same and the tests all pass...